# Optimize XxHash3 on ARM platform